+++ ui/components/src/Table/useVirtualizedTableKeyboardNav.tsx
@@ -1,7 +1,8 @@
 import { useCallback, useState } from 'react';
 import { TableVirtuosoHandle } from 'react-virtuoso';
+import { useTableKeyboardNav, UseTableKeyboardNavProps } from './useTableKeyboardNav';
 
-type UseTableKeyboardNavProps = {
+interface UseVirtualizedTableKeyboardNavProps extends Omit<UseTableKeyboardNavProps, 'onActiveCellChange'> {
   visibleRange: React.MutableRefObject<{
     startIndex: number;
     endIndex: number;
@@ -9,134 +10,90 @@ type UseTableKeyboardNavProps = {
   virtualTable: React.RefObject<TableVirtuosoHandle>;
   maxRows: number;
   maxColumns: number;
-};
-
-type TableCellPosition = {
-  row: number;
-  column: number;
-};
-
-const DEFAULT_ACTIVE_CELL: TableCellPosition = {
-  row: 0,
-  column: 0,
-};
-
-const ARROW_KEYS = ['ArrowRight', 'ArrowLeft', 'ArrowUp', 'ArrowDown'];
-
-function isArrowKey(key: string) {
-  return ARROW_KEYS.includes(key);
 }
 
+/**
+ * Hook for managing keyboard navigation when using a virtualized table.
+ */
 export function useVirtualizedTableKeyboardNav({
   visibleRange,
   virtualTable,
   maxRows,
   maxColumns,
-}: UseTableKeyboardNavProps) {
-  console.log('virt table hook');
-  const [activeCell, setActiveCell] = useState<TableCellPosition>(DEFAULT_ACTIVE_CELL);
-  const [isActive, setIsActive] = useState(false);
-
-  const handleCellFocus = (cellPosition: TableCellPosition) => {
-    if (cellPosition.column === activeCell.column && cellPosition.row === activeCell.row && isActive) {
-      return;
-    }
-    setIsActive(true);
-  };
-
-  const handleKeyDown: React.KeyboardEventHandler<HTMLTableElement> = useCallback(
-    (e) => {
-      console.log('__keydown__');
-      console.log(visibleRange.current);
-      // Including some of the basic a11y keyboard interaction patterns from:
-      // https://www.w3.org/WAI/ARIA/apg/patterns/grid/
-      // TODO: add other keyboard combos.
+}: UseVirtualizedTableKeyboardNavProps) {
+  const baseKeyboard = useTableKeyboardNav({
+    maxRows,
+    maxColumns,
+    onActiveCellChange: (e, currentPosition, defaultNewPosition) => {
       const key = e.key;
 
-      if (isArrowKey(key) || key === 'Home' || key === 'End' || key === 'PageDown' || key === 'PageUp') {
-        setActiveCell((curActiveCell) => {
-          let nextRow: number = curActiveCell.row;
-          let nextColumn: number = curActiveCell.column;
+      const defaultValueChanged =
+        defaultNewPosition &&
+        (currentPosition.column !== defaultNewPosition.column || currentPosition.row !== defaultNewPosition.row);
+      const nextRow = defaultNewPosition?.row ?? currentPosition.row;
 
-          if (key === 'ArrowRight' && nextColumn < maxColumns - 1) {
-            nextColumn += 1;
-          } else if (key === 'ArrowLeft' && nextColumn > 0) {
-            nextColumn -= 1;
-          } else if (key === 'ArrowDown' && nextRow < maxRows - 1) {
-            e.preventDefault();
-            nextRow += 1;
-
-            // TODO: Only do when needed
-            if (nextRow - 1 < visibleRange.current.startIndex || nextRow - 1 > visibleRange.current.endIndex) {
-              virtualTable.current?.scrollToIndex({
-                index: nextRow - 1,
-                align: 'end',
-              });
-            }
-          } else if (key === 'ArrowUp' && nextRow > 0) {
-            e.preventDefault();
-            nextRow -= 1;
-
-            // TODO: Only do when needed
-            if (nextRow - 1 < visibleRange.current.startIndex || nextRow - 1 > visibleRange.current.endIndex) {
-              virtualTable.current?.scrollToIndex({
-                index: nextRow - 1,
-                align: 'start',
-              });
-            }
-          } else if (key === 'Home') {
-            nextRow = 0;
-            nextColumn = 0;
-            virtualTable.current?.scrollToIndex({
-              index: nextRow - 1,
-              align: 'start',
-            });
-          } else if (key === 'End') {
-            nextRow = maxRows - 1;
-            nextColumn = maxColumns - 1;
-            virtualTable.current?.scrollToIndex({
-              index: nextRow - 1,
-              align: 'start',
-            });
-          } else if (key === 'PageDown') {
-            console.log('page down');
-            e.preventDefault();
-            // Add 1 to account for header
-            // console.log(maxRows - 1, visibleRange.endIndex + 1);
-            nextRow = Math.min(maxRows - 1, visibleRange.current.endIndex + 1);
-            // console.log(nextRow);
-            virtualTable.current?.scrollToIndex({
-              index: nextRow - 1,
-              align: 'start',
-            });
-          } else if (key === 'PageUp') {
-            console.log('page up');
-            e.preventDefault();
-            // Minus 1 to account for header
-            nextRow = Math.max(0, visibleRange.current.startIndex - 1);
-            virtualTable.current?.scrollToIndex({
-              index: nextRow - 1,
-              align: 'end',
-            });
-          }
-
-          if (nextRow === curActiveCell.row && nextColumn === curActiveCell.column) {
-            // Return original to avoid creating a new object if nothing
-            // changed.
-            return curActiveCell;
-          }
-
-          return { column: nextColumn, row: nextRow };
+      if (key === 'ArrowDown' && defaultValueChanged) {
+        // Use default cell position. Shift the virtual table scroll position
+        // when needed to make the active cell visible.
+        if (nextRow - 1 < visibleRange.current.startIndex || nextRow - 1 > visibleRange.current.endIndex) {
+          virtualTable.current?.scrollToIndex({
+            index: nextRow - 1,
+            align: 'end',
+          });
+        }
+      } else if (key === 'ArrowUp' && defaultValueChanged) {
+        // Use default cell position. Shift the virtual table scroll position
+        // when needed to make the active cell visible.
+        if (nextRow - 1 < visibleRange.current.startIndex || nextRow - 1 > visibleRange.current.endIndex) {
+          virtualTable.current?.scrollToIndex({
+            index: nextRow - 1,
+            align: 'start',
+          });
+        }
+      } else if (defaultValueChanged && (key === 'Home' || key === 'End')) {
+        // Use default cell position. Shift the virtual table scroll position
+        // when needed to make the active cell visible.
+        virtualTable.current?.scrollToIndex({
+          index: Math.max(nextRow - 1, 0),
+          align: 'start',
         });
-      }
-    },
-    [maxColumns, maxRows, virtualTable, visibleRange]
-  );
+      } else if (key === 'PageDown') {
+        // Full handling of logic for `PageDown` because there is no default.
+        e.preventDefault();
 
-  return {
-    activeCell,
-    isActive,
-    onTableKeyDown: handleKeyDown,
-    onCellFocus: handleCellFocus,
-  };
+        let nextRow = currentPosition.row;
+        // Add 1 to account for header
+
+        nextRow = Math.min(maxRows - 1, visibleRange.current.endIndex + 1);
+
+        virtualTable.current?.scrollToIndex({
+          index: nextRow - 1,
+          align: 'start',
+        });
+
+        return {
+          row: nextRow,
+          column: currentPosition.column,
+        };
+      } else if (key === 'PageUp') {
+        // Full handling of logic for `PageUp` because there is no default.
+        let nextRow = currentPosition.row;
+        // Minus 1 to account for header
+        nextRow = Math.max(0, visibleRange.current.startIndex - 1);
+        virtualTable.current?.scrollToIndex({
+          index: nextRow - 1,
+          align: 'end',
+        });
+
+        return {
+          row: nextRow,
+          column: currentPosition.column,
+        };
+      }
+
+      return defaultNewPosition;
+    },
+  });
+
+  return baseKeyboard;
 }
